update for ouster
1. update point cloud type for ouster lidar
2. update parameters for ouster lidar

+++ include/lvt2calib/ouster_utils.h
@@ -77,7 +77,7 @@ namespace Ouster {
     float intensity;
     uint32_t t;
     uint16_t reflectivity;
-    uint8_t ring;
+    uint16_t ring;
     uint16_t ambient;
     float range;
     EIGEN_MAKE_ALIGNED_OPERATOR_NEW
@@ -152,7 +152,7 @@ POINT_CLOUD_REGISTER_POINT_STRUCT(Ouster::Point,
     // use std::uint32_t to avoid conflicting with pcl::uint32_t
     (std::uint32_t, t, t)
     (std::uint16_t, reflectivity, reflectivity)
-    (std::uint8_t, ring, ring)
+    (std::uint16_t, ring, ring)
     (std::uint16_t, ambient, ambient)
     (float, range, range)
 )
